fastapi_server/pyproject.toml
[project]
name = "fastapi-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "jinja2>=3.1.6",
    "numpy>=2.4.0",
    "pandas>=2.2.0",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
    "requests>=2.32.5",
    "sqlmodel>=0.0.22",
    "uvicorn>=0.40.0",
]
